+++ catalog/app/containers/Bucket/PackageDialog/State/index.tsx
@@ -1,3 +1,3 @@
-// Re-export all state-related exports from separate files
 export * from './State'
 export * from './Provider'
+export { isPackageHandle } from './manifest'
